+++ src/main/scala/com/yogpc/qp/render/Box.scala
@@ -379,7 +379,7 @@ object Box {
 
   private def getPosVector(x: Float, y: Float, z: Float, matrix: MatrixStack): Vector4f = {
     val vec3i = Direction.UP.getDirectionVec
-    vector3f.set(vec3i.getX, vec3i.getY, vec3i.getZ)
+    vector3f.set(vec3i.getX.toFloat, vec3i.getY.toFloat, vec3i.getZ.toFloat)
     val matrix4f = matrix.getLast.getMatrix
     vector3f.transform(matrix.getLast.getNormal)
